fix RTC, add set RTC, modified i2c.c

+++ Inc/hih.h
@@ -1,6 +1,6 @@
 #ifndef __HIH_H
 #define __HIH_H
 
-uint16_t mVToRH(int16_t Vadc, int16_t spRH, int16_t pvT0);
+uint8_t mVToRH(int16_t Vadc, int16_t spRH, int16_t pvT0);
 
 #endif /* __HIH_H */
